requirements.txt
huggingface-hub==0.30.2
matplotlib==3.10.1
mpmath==1.3.0
numpy==2.2.5
pillow==11.2.1
torch==2.6.0
transformers==4.51.3

